+++ qupath-experimental/src/main/java/qupath/experimental/commands/SimpleThresholdCommand.java
@@ -20,6 +20,7 @@ import javafx.scene.control.Label;
 import javafx.scene.control.Spinner;
 import javafx.scene.control.SpinnerValueFactory;
 import javafx.scene.layout.GridPane;
+import javafx.scene.layout.Priority;
 import javafx.stage.Stage;
 import qupath.lib.classifiers.pixel.PixelClassificationImageServer;
 import qupath.lib.classifiers.pixel.PixelClassifier;
@@ -72,6 +73,7 @@ public class SimpleThresholdCommand implements PathCommand {
 	public void run() {
 		if (qupath.getImageData() == null) {
 			Dialogs.showNoImageError("Simple threshold");
+			return;
 		}
 		if (stage == null || !stage.isShowing())
 			showGUI();
@@ -234,9 +236,9 @@ public class SimpleThresholdCommand implements PathCommand {
 		PaneTools.setFillWidth(Boolean.TRUE, comboResolutions, comboPrefilter,
 				transforms, spinner, sigmaSpinner, classificationsAbove, classificationsEqual, classificationsBelow,
 				btnSave, btnClassifyObjects, btnCreateObjects, tilePane);
-//		PaneToolsFX.setHGrowPriority(Priority.ALWAYS, comboResolutions, comboPrefilter,
-//				transforms, spinner, sigmaSpinner, classificationsAbove, classificationsEqual, classificationsBelow,
-//				btnSave, btnClassifyObjects, btnCreateObjects, tilePane);
+		PaneTools.setHGrowPriority(Priority.ALWAYS, comboResolutions, comboPrefilter,
+				transforms, spinner, sigmaSpinner, classificationsAbove, classificationsEqual, classificationsBelow,
+				btnSave, btnClassifyObjects, btnCreateObjects, tilePane);
 		
 		updateGUI();
 		
@@ -252,12 +254,14 @@ public class SimpleThresholdCommand implements PathCommand {
 		stage.setResizable(false);
 		stage.show();
 		
-		stage.sizeToScene();
+//		stage.sizeToScene();
 		
 		stage.setOnHiding(e -> {
 			for (var entry : map.entrySet()) {
 				if (entry.getKey().getCustomPixelLayerOverlay() == entry.getValue()) {
-					PixelClassificationImageServer.setPixelLayer(entry.getKey().getImageData(), null);
+					var imageData = entry.getKey().getImageData();
+					if (imageData != null)
+						PixelClassificationImageServer.setPixelLayer(entry.getKey().getImageData(), null);
 					selectedOverlay.set(null);
 					entry.getKey().resetCustomPixelLayerOverlay();
 				}