+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This is a *milestone* (i.e. still in development) version made available to try 
   * Updated to Groovy 3 - scripts now support more recent Java syntax (e.g. lambdas, try-with-resources)
   * Pasting files results in them being converted to absolute paths
   * New 'Paste & escape' command to automatically escape characters for Java Strings
+* New 'Measure -> Export measurements' command to export measurements for multiple images within a project
 * Scriptable 'Select objects by classification' command
 * Optionally show/hide annotation names in the viewer (shortcut key 'N')
 * Updated methods to save/load points within the counting tool